added defining fields
